Convert test report integration test to new junit multiversion test
